Refactoring CLI: move modify related parts to modify.go
Changelog: none
